add a LICENSE file containing the MIT license text, as is necessary
As far as I can discern Paul's intent is that all of this code is licensed as
MIT, but that license requires a copy of the license text to be provided with
the source. As ogr2osm-translations can be cloned separately from ogr2osm,
it's probably a good idea to have a copy of LICENSE here. I checked the commit
history of each translation file and AFAICS only Paul has contributed, so the
attribution should be correct (assuming my Vancouver translation is accepted).